# api_contract/openapi.json
{
  "components": {
    "schemas": {
      "CategoryListItem": {
        "properties": {
          "id": {
            "description": "Stable category identifier",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "description": "Display name of the category",
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "CategoryListItem",
        "type": "object"
      },
      "DataOverview": {
        "properties": {
          "anomalie": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anomalie"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "umap_x": {
            "title": "Umap X",
            "type": "number"
          },
          "umap_y": {
            "title": "Umap Y",
            "type": "number"
          },
          "umap_z": {
            "title": "Umap Z",
            "type": "number"
          },
          "uuid": {
            "title": "Uuid",
            "type": "string"
          }
        },
        "required": [
          "uuid",
          "umap_x",
          "umap_y",
          "umap_z",
          "label",
          "category",
          "filename"
        ],
        "title": "DataOverview",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LabeledSample": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "uuid": {
            "title": "Uuid",
            "type": "string"
          }
        },
        "required": [
          "uuid",
          "label",
          "category"
        ],
        "title": "LabeledSample",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Backend API for the Audioexplorer application.",
    "title": "Audioexplorer API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/sounds/categories": {
      "get": {
        "operationId": "get_category_list_api_v1_sounds_categories_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CategoryListItem"
                  },
                  "title": "Response Get Category List Api V1 Sounds Categories Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Category List",
        "tags": [
          "sounds"
        ]
      }
    },
    "/api/v1/sounds/categories/{category_id}": {
      "get": {
        "operationId": "get_category_by_id_api_v1_sounds_categories__category_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "category_id",
            "required": true,
            "schema": {
              "title": "Category Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryListItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Category By Id",
        "tags": [
          "sounds"
        ]
      }
    },
    "/api/v1/sounds/labeled-samples": {
      "get": {
        "operationId": "get_labeled_samples_api_v1_sounds_labeled_samples_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LabeledSample"
                  },
                  "title": "Response Get Labeled Samples Api V1 Sounds Labeled Samples Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Labeled Samples",
        "tags": [
          "sounds"
        ]
      }
    },
    "/api/v1/sounds/labeled-samples/{uuid}": {
      "get": {
        "operationId": "get_labeled_sample_by_uuid_api_v1_sounds_labeled_samples__uuid__get",
        "parameters": [
          {
            "in": "path",
            "name": "uuid",
            "required": true,
            "schema": {
              "title": "Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LabeledSample"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Labeled Sample By Uuid",
        "tags": [
          "sounds"
        ]
      }
    },
    "/api/v1/sounds/overviews": {
      "get": {
        "operationId": "get_all_data_overviews_api_v1_sounds_overviews_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DataOverview"
                  },
                  "title": "Response Get All Data Overviews Api V1 Sounds Overviews Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get All Data Overviews",
        "tags": [
          "sounds"
        ]
      }
    },
    "/api/v1/sounds/overviews/{uuid}": {
      "get": {
        "operationId": "get_data_overview_by_uuid_api_v1_sounds_overviews__uuid__get",
        "parameters": [
          {
            "in": "path",
            "name": "uuid",
            "required": true,
            "schema": {
              "title": "Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataOverview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Data Overview By Uuid",
        "tags": [
          "sounds"
        ]
      }
    }
  }
}